[Intro] Robbie: “Are you, uh, a real villain?” Bobbie: “Well, uh, technically... nah.” Robbie: “Have you ever caught a good guy, like, uh, like a real superhero?” Bobbie: “Nah.” Robbie: “Have you ever tried a disguise?” Bobbie: “Nah, nah...” Robbie: “Alright! I can see that I will have to teach you how to be villains!”
[Hook]
Hey!
We are Number One
Hey!
We are Number One

[Verse 1]
Now listen closely
Here's a little lesson in trickery
This is going down in history
If you wanna be a Villain Number One
You have to chase a superhero on the run
Just follow my moves, and sneak around
Be careful not to make a sound
(Shh)
(No, don't touch that!)

[Hook]
We are Number One
Hey!
We are Number One
We are Number One